% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complement.r
\name{complement}
\alias{complement}
\title{Returns the base pair complement of a feature file.}
\usage{
complement(i, g)
}
\arguments{
\item{i}{<bed/gff/vcf>}

\item{g}{<genome>}
}
\description{
Returns the base pair complement of a feature file.
}
